# src/app/api/openapi.json/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Preca API",
    "description": "API para la gestión de servicios de precalificación y screening en México",
    "version": "1.0.0",
    "contact": {
      "name": "Preca Support",
      "email": "soporte@preca.mx"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    }
  ],
...
  "paths": {
...
    "/api-keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "Listar API keys",
        "description": "Obtiene todas las API keys del usuario autenticado. Requiere autenticación.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "activeOnly",
            "in": "query",
            "description": "Filtrar solo claves activas",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de API keys obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKey"
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Crear API key",
        "description": "Crea una nueva API key con permisos específicos. La clave completa solo se muestra una vez. Requiere autenticación.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyWithRawKey"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Límite de API keys excedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api-keys/{id}": {
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Eliminar API key",
        "description": "Elimina permanentemente una API key. Esta acción no se puede deshacer. Requiere autenticación.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la API key (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "API key deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado - No puedes eliminar claves de otros usuarios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "API key no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/files/upload": {
      "post": {
        "tags": [
          "Archivos"
        ],
        "summary": "Subir archivo",
        "description": "Sube un archivo (imagen o PDF) con compresión automática para imágenes. El archivo se almacena en Cloudflare R2. Requiere autenticación.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo a subir (imagen: JPEG, PNG, GIF, WEBP o PDF). Las imágenes se comprimen automáticamente a JPEG con calidad 85%. Máximo 10MB."
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archivo subido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fileUrl": {
                      "type": "string",
                      "description": "URL del archivo subido (válido por 7 días)",
                      "example": "https://6809b8cd12e8e1f392555597f0aa0ca7.r2.cloudflarestorage.com/..."
                    },
                    "fileKey": {
                      "type": "string",
                      "description": "Clave del archivo en el almacenamiento",
                      "example": "form-uploads/1734567890123-documento.pdf"
                    },
                    "fileName": {
                      "type": "string",
                      "description": "Nombre del archivo original",
                      "example": "documento.pdf"
                    },
                    "contentType": {
                      "type": "string",
                      "description": "Tipo MIME del archivo (image/jpeg para imágenes comprimidas)",
                      "example": "image/jpeg"
                    },
                    "size": {
                      "type": "integer",
                      "description": "Tamaño del archivo en bytes (después de compresión si aplica)",
                      "example": 245678
                    },
                    "uploadedAt": {
                      "type": "string",
                      "format": "date-time",
                      "description": "Fecha de carga"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Archivo inválido (tipo no permitido, tamaño excedido, o no se proporcionó archivo)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "examples": {
                  "noFile": {
                    "summary": "No se proporcionó archivo",
                    "value": {
                      "error": "No file provided"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado - API key o sesión inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor o error de almacenamiento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/coupons/{id}": {
      "delete": {
        "tags": [
          "Cupones"
        ],
        "summary": "Eliminar cupón",
        "description": "Elimina un cupón de la base de datos y de Stripe. Solo administradores. Requiere autenticación.",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del cupón (UUID)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cupón eliminado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Coupon deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "No autorizado - Requiere rol de administrador",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Cupón no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "API Key de autenticación en formato: preca_xxxx_xxxxxxxxxxxxxxxx"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Cookie de sesión para usuarios autenticados vía web"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensaje de error"
          }
        },
        "required": [
          "error"
        ]
      },
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "ID único de la API key"
          },
          "keyPrefix": {
            "type": "string",
            "description": "Prefijo de la clave (primeros 10 caracteres)",
            "example": "preca_a1b2"
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Permisos asignados a la clave",
            "example": [
              "services:read",
              "screenings:write"
            ]
          },
          "isActive": {
            "type": "boolean",
            "description": "Indica si la clave está activa"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha del último uso"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de expiración"
          }
        },
        "required": [
          "id",
          "keyPrefix",
          "scopes",
          "isActive",
          "createdAt"
        ]
      },
      "ApiKeyWithRawKey": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/ApiKey"
          }
        ],
        "properties": {
          "rawKey": {
            "type": "string",
            "description": "La clave completa (solo se muestra al crear)",
            "example": "preca_a1b2c3d4e5f6g7h8i9j0k1l2m3n4o5p6q7r8s9t0u1v2w3x4y5z6"
          }
        }
      },
      "CreateApiKeyRequest": {
        "type": "object",
        "properties": {
          "scopes": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "services:read",
                "services:write",
                "screenings:read",
                "screenings:write",
                "users:read"
              ]
            },
            "description": "Permisos para la API key",
            "minItems": 1,
            "example": [
              "services:read",
              "screenings:write"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Fecha de expiración opcional"
          }
        },
        "required": [
          "scopes"
        ]
      },
...
    }
  }
}